Update infrared.md to mention ESP8266 quirks
Short chains (8 WS2812b's) work fine with bit-banging + IR, longer ones (20+) make IR reception unreliable. It's easy to reproduce if you connect an IR receiver to a ESP8266 devkit and check the console output, it doesn't need a LED chain to reproduce. This is not an issue on ESP32 (as far as I could tell, maybe the processing speed is just fast enough there). Unreliable IR reception is super-annoying to debug and moving the chain from GPIO4 to GPIO1,2, or 3 was a quick fix I didn't expect, and which was non-trivial to figure out. Older Wled versions throw a warning when compiling with bit-banging pins, that's when it clicked.