Fix formatting on compatible-controllers
There's a section on the [compatible
controllers](https://kno.wled.ge/basics/compatible-controllers/) page
that's clearly intended to be a numbered list but gets rendered as a
single paragraph because it's not formatted as a markdown list.
Indenting the list makes markdown recognize it as a numbered list and
it's much easier to read when rendered!